requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
audioread==3.0.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1702571698061/work
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
crepe==0.0.15
croniter==1.3.15
cvxopt==1.3.2
cycler==0.11.0
dateutils==0.6.12
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1649586340600/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepdiff==6.7.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
enum34==1.1.10
exceptiongroup==1.2.1
fastapi==0.88.0
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h11==0.14.0
h5py==3.8.0
hmmlearn==0.3.0
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
inquirer==2.10.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1666723258080/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240553635/work
itsdangerous==2.1.2
jams==0.3.4
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.17.3
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1658332345782/work
keras==2.11.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==1.9.5
lightning-cloud==0.5.69
lightning-utilities==0.10.1
line-profiler==4.1.3
llvmlite==0.39.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mir-eval==0.7
mlconfig==0.1.8
msaf @ file:///home/longshen/work/SoloDetection/svm_baseline/msaf
msgpack==1.0.5
multidict==6.0.5
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
oyaml==1.0
packaging==24.0
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.2
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==3.19.6
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.15
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-multipart==0.0.8
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1652965483789/work
readchar==4.0.5
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.2.2
rich==13.7.1
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
starlette==0.22.0
starsessions==1.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torchmetrics==0.11.4
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827244717/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.18
uvicorn==0.22.0
vmo==0.30.5
wcwidth==0.2.13
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
